Hash, Map: make literals create Map instead of Hash objects
I accidentally did this in the bytecode compiler, and Map seems better than Hash in general.

This uncovered a serious bug in Map, where a Map object created in a different file than the one where it is used
woud lead to a segfault. I changed a bunch of functions that took EHI objects to take EHInterpreter instead in an
attempt to fix this bug, but decided that that solution was still flawed. For now I worked around the bug with
some mutable state in the Map object, but bigger changes to the EHI/EHInterpreter system are ultimately needed.